requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.47.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
environ==1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
psycopg2==2.9.2
Py-Mono==0.0.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
